+++ packages/elixir-client/lib/electric/client/application.ex
@@ -8,7 +8,8 @@ defmodule Electric.Client.Application do
     children = [
       {Registry, name: Electric.Client.Registry, keys: :unique},
       {Finch, name: Electric.Client.Finch},
-      {DynamicSupervisor, name: Electric.Client.RequestSupervisor, strategy: :one_for_one}
+      {DynamicSupervisor, name: Electric.Client.RequestSupervisor, strategy: :one_for_one},
+      Electric.Client.ExpiredShapesCache
     ]
 
     Supervisor.start_link(children, strategy: :one_for_one, name: __MODULE__)