+++ packages/sync-service/lib/electric/stack_supervisor/telemetry.ex
@@ -60,7 +60,8 @@ defmodule Electric.StackSupervisor.Telemetry do
       [
         {__MODULE__, :count_shapes, [stack_id]},
         {__MODULE__, :report_write_buffer_size, [stack_id]},
-        {__MODULE__, :report_retained_wal_size, [stack_id, config.replication_opts[:slot_name]]}
+        {__MODULE__, :report_retained_wal_size, [stack_id, config.replication_opts[:slot_name]]},
+        {__MODULE__, :report_disk_usage, [stack_id]}
       ]
     end
 
@@ -158,11 +159,9 @@ defmodule Electric.StackSupervisor.Telemetry do
       case ElectricTelemetry.DiskUsage.current(stack_id) do
         {:ok, usage_bytes, measurement_duration} ->
           Electric.Telemetry.OpenTelemetry.execute(
-            [:electric, :storage],
-            %{used: usage_bytes, measurement_duration_ms: measurement_duration},
-            %{
-              stack_id: stack_id
-            }
+            [:electric, :storage, :used],
+            %{bytes: usage_bytes, measurement_duration: measurement_duration},
+            %{stack_id: stack_id}
           )
 
         :pending ->